Refactor model
