+++ app/src/main/java/com/example/colornote/model/CheckList.java
@@ -2,59 +2,5 @@ package com.example.colornote.model;
 
 import java.sql.Date;
 
-public class CheckList {
-    private int id;
-    private String title;
-    private int colorId;
-    private Date reminder;
-    private Date modifiedDate;
-    private int status;
-
-    public int getColorId() {
-        return colorId;
-    }
-
-    public void setColorId(int colorId) {
-        this.colorId = colorId;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public Date getReminder() {
-        return reminder;
-    }
-
-    public void setReminder(Date reminder) {
-        this.reminder = reminder;
-    }
-
-    public Date getModifiedDate() {
-        return modifiedDate;
-    }
-
-    public void setModifiedDate(Date modifiedDate) {
-        this.modifiedDate = modifiedDate;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public void setStatus(int status) {
-        this.status = status;
-    }
+public class CheckList extends Task{
 }
