+++ app/src/main/java/com/example/colornote/model/Text.java
@@ -2,38 +2,8 @@ package com.example.colornote.model;
 
 import java.util.Date;
 
-public class Text {
-    private int id;
-    private String title;
+public class Text extends Task{
     private String content;
-    private int colorId;
-    private Date reminder;
-    private Date modifiedDate;
-    private int status;
-
-    public long getColorId() {
-        return colorId;
-    }
-
-    public void setColorId(int colorId) {
-        this.colorId = colorId;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
 
     public String getContent() {
         return content;
@@ -42,30 +12,4 @@ public class Text {
     public void setContent(String content) {
         this.content = content;
     }
-
-    public Date getReminder() {
-        return reminder;
-    }
-
-    public void setReminder(Date reminder) {
-        this.reminder = reminder;
-    }
-
-    public Date getModifiedDate() {
-        return modifiedDate;
-    }
-
-    public void setModifiedDate(Date modifiedDate) {
-        this.modifiedDate = modifiedDate;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public void setStatus(int status) {
-        this.status = status;
-    }
-
-
 }
